Realloc Optimization for GBTLB allocator
HBW API has one interface hbw_posix_memalign_psize to allocate
memory with Gigabyte pages. But Realloc to follow is a completely
legal usage of the API. Currently realloc was not supported as
it would just increase the memory allocated. But if there is
enough memory, realloc would return the same pointer which is
mapped to the 1GB memory, if not realloc would allocate more memory
and return a new pointer. This avoids a mmap and mbind call,
everytime there is memory available with the same 1GB page and
realloc is invoked.