autohbw: move out from examples to autohbw subdir
There was a question from customer why autohbw is not shipped with RPMS.
It was not shipped as autohbw was part of examples which we do not ship
at all. But, turns out that autohbw is widely promoted in HPC world,
thus we need to ship autohbw. To not introduce an exception, decision
was made to move autohbw from examples (which we still do not package
into RPMS) to separate subdirectory and package autohbw along with
libmemkind.

Change-Id: Ifbdb95ea2bc4f0a4fb09d717b8825b3555638e5e
Signed-off-by: Lukasz Anaczkowski <lukasz.anaczkowski@intel.com>